requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
distlib==0.3.1
Django
django-cors-headers
django-rest-knox
djangorestframework
djangorestframework-simplejwt
djongo
dnspython==2.0.0
drf-yasg==1.17.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
isort==5.5.3
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-client==0.0.1
mysqlclient==2.0.1
packaging==20.4
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pytz==2020.1
razorpay==1.2.0
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
wrapt==1.12.1
